ex1

+++ IP/Exercicios-Aula/Ficha_n5/ex1.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int squareRoot(float n)
+float squareRoot(float n)
 {
     return n * n;
 }
